add disk_get_part_count API

+++ FAT32/common/common_types.h
@@ -4,9 +4,9 @@
 #include <stdint.h>
 
 typedef uint8_t u8_t;
-typedef uint8_t u16_t;
-typedef uint8_t u32_t;
-typedef uint8_t u64_t;
+typedef uint16_t u16_t;
+typedef uint32_t u32_t;
+typedef uint64_t u64_t;
 
 typedef enum _efat32_err_t {
     FAT32_ERR_OK = 0,
